requirements.txt
PyExecJS
requests